requirements.txt
PyExecJS
requests